// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const UserRole = {
  TOURIST: 'TOURIST',
  GUIDE: 'GUIDE',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TourCategory = {
  FOOD: 'FOOD',
  ART: 'ART',
  ADVENTURE: 'ADVENTURE',
  HISTORY: 'HISTORY',
  NIGHTLIFE: 'NIGHTLIFE',
  NATURE: 'NATURE',
  WILDLIFE: 'WILDLIFE',
  SHOPPING: 'SHOPPING',
  HERITAGE: 'HERITAGE',
  OTHER: 'OTHER'
} as const

export type TourCategory = (typeof TourCategory)[keyof typeof TourCategory]


export const BookingStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const PaymentMethod = {
  STRIPE: 'STRIPE',
  SSL_COMMERZ: 'SSL_COMMERZ',
  OTHER: 'OTHER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]